Break up solve/solve into smaller modules
