+++ crates/compiler/late_solve/src/lib.rs
@@ -12,8 +12,8 @@ use roc_error_macros::internal_error;
 use roc_module::symbol::ModuleId;
 use roc_module::symbol::Symbol;
 use roc_solve::ability::AbilityResolver;
-use roc_solve::solve::Pools;
 use roc_solve::specialize::{compact_lambda_sets_of_vars, DerivedEnv, Phase};
+use roc_solve::Pools;
 use roc_types::subs::{get_member_lambda_sets_at_region, Content, FlatType, LambdaSet};
 use roc_types::subs::{ExposedTypesStorageSubs, Subs, Variable};
 use roc_types::types::Polarity;
