+++ crates/compiler/solve/src/pools.rs
@@ -1,0 +1,59 @@
+use roc_types::subs::{Rank, Variable};
+
+const DEFAULT_POOLS: usize = 8;
+
+#[derive(Clone, Debug)]
+pub struct Pools(Vec<Vec<Variable>>);
+
+impl Default for Pools {
+    fn default() -> Self {
+        Pools::new(DEFAULT_POOLS)
+    }
+}
+
+impl Pools {
+    pub fn new(num_pools: usize) -> Self {
+        Pools(vec![Vec::new(); num_pools])
+    }
+
+    pub fn len(&self) -> usize {
+        self.0.len()
+    }
+
+    pub fn is_empty(&self) -> bool {
+        self.0.is_empty()
+    }
+
+    pub fn get_mut(&mut self, rank: Rank) -> &mut Vec<Variable> {
+        match self.0.get_mut(rank.into_usize()) {
+            Some(reference) => reference,
+            None => panic!("Compiler bug: could not find pool at rank {}", rank),
+        }
+    }
+
+    pub fn get(&self, rank: Rank) -> &Vec<Variable> {
+        match self.0.get(rank.into_usize()) {
+            Some(reference) => reference,
+            None => panic!("Compiler bug: could not find pool at rank {}", rank),
+        }
+    }
+
+    pub fn iter(&self) -> std::slice::Iter<'_, Vec<Variable>> {
+        self.0.iter()
+    }
+
+    pub fn split_last(mut self) -> (Vec<Variable>, Vec<Vec<Variable>>) {
+        let last = self
+            .0
+            .pop()
+            .unwrap_or_else(|| panic!("Attempted to split_last() on non-empty Pools"));
+
+        (last, self.0)
+    }
+
+    pub fn extend_to(&mut self, n: usize) {
+        for _ in self.len()..n {
+            self.0.push(Vec::new());
+        }
+    }
+}
