+++ crates/compiler/solve/src/solve.rs
@@ -4,36 +4,32 @@ use crate::ability::{
     resolve_ability_specialization, type_implementing_specialization, AbilityImplError,
     CheckedDerives, ObligationCache, PendingDerivesTable, Resolved,
 };
+use crate::deep_copy::deep_copy_var_in;
 use crate::module::{SolveCtx, Solved};
+use crate::pools::Pools;
 use crate::specialize::{
     compact_lambda_sets_of_vars, AwaitingSpecializations, CompactionResult, DerivedEnv, SolvePhase,
 };
+use crate::to_var::{either_type_index_to_var, type_to_var};
+use crate::Aliases;
 use bumpalo::Bump;
 use roc_can::abilities::{AbilitiesStore, MemberSpecializationInfo};
 use roc_can::constraint::Constraint::{self, *};
-use roc_can::constraint::{Constraints, Cycle, LetConstraint, OpportunisticResolve, TypeOrVar};
+use roc_can::constraint::{Constraints, Cycle, LetConstraint, OpportunisticResolve};
 use roc_can::expected::{Expected, PExpected};
-use roc_collections::all::MutMap;
-use roc_collections::soa::{Index, Slice};
 use roc_debug_flags::dbg_do;
 #[cfg(debug_assertions)]
 use roc_debug_flags::ROC_VERIFY_RIGID_LET_GENERALIZED;
 use roc_error_macros::internal_error;
-use roc_module::ident::TagName;
 use roc_module::symbol::Symbol;
 use roc_problem::can::CycleEntry;
 use roc_region::all::Loc;
 use roc_solve_problem::TypeError;
 use roc_types::subs::{
-    self, AliasVariables, Content, Descriptor, FlatType, GetSubsSlice, LambdaSet, Mark,
-    OptVariable, Rank, RecordFields, Subs, SubsSlice, TagExt, TupleElems, UlsOfVar, UnionLabels,
-    UnionLambdas, UnionTags, Variable, VariableSubsSlice,
+    self, Content, Descriptor, FlatType, GetSubsSlice, Mark, OptVariable, Rank, Subs, TagExt,
+    UlsOfVar, Variable,
 };
-use roc_types::types::{
-    gather_fields_unsorted_iter, gather_tuple_elems_unsorted_iter, AliasKind, AliasShared,
-    Category, ExtImplicitOpenness, OptAbleVar, Polarity, Reason, RecordField, Type, TypeExtension,
-    TypeTag, Types, Uls,
-};
+use roc_types::types::{Category, Polarity, Reason, RecordField, Type, TypeExtension, Types, Uls};
 use roc_unify::unify::{
     unify, unify_introduced_ability_specialization, Env as UEnv, Mode, Obligated,
     SpecializationLsetCollector, Unified::*,
@@ -91,395 +87,6 @@ pub use scope::Scope;
 // Thus instead the inferred type for `id` is generalized (see the `generalize` function) to `a -> a`.
 // Ranks are used to limit the number of type variables considered for generalization. Only those inside
 // of the let (so those used in inferring the type of `\x -> x`) are considered.
-
-use roc_types::types::Alias;
-
-#[derive(Debug, Clone, Copy)]
-struct DelayedAliasVariables {
-    start: u32,
-    type_variables_len: u8,
-    lambda_set_variables_len: u8,
-    recursion_variables_len: u8,
-    infer_ext_in_output_variables_len: u8,
-}
-
-impl DelayedAliasVariables {
-    fn recursion_variables(self, variables: &mut [OptAbleVar]) -> &mut [OptAbleVar] {
-        let start = self.start as usize
-            + (self.type_variables_len + self.lambda_set_variables_len) as usize;
-        let length = self.recursion_variables_len as usize;
-
-        &mut variables[start..][..length]
-    }
-
-    fn lambda_set_variables(self, variables: &mut [OptAbleVar]) -> &mut [OptAbleVar] {
-        let start = self.start as usize + self.type_variables_len as usize;
-        let length = self.lambda_set_variables_len as usize;
-
-        &mut variables[start..][..length]
-    }
-
-    fn type_variables(self, variables: &mut [OptAbleVar]) -> &mut [OptAbleVar] {
-        let start = self.start as usize;
-        let length = self.type_variables_len as usize;
-
-        &mut variables[start..][..length]
-    }
-
-    fn infer_ext_in_output_variables(self, variables: &mut [OptAbleVar]) -> &mut [OptAbleVar] {
-        let start = self.start as usize
-            + (self.type_variables_len
-                + self.lambda_set_variables_len
-                + self.recursion_variables_len) as usize;
-        let length = self.infer_ext_in_output_variables_len as usize;
-
-        &mut variables[start..][..length]
-    }
-}
-
-#[derive(Debug, Default)]
-pub struct Aliases {
-    aliases: Vec<(Symbol, Index<TypeTag>, DelayedAliasVariables, AliasKind)>,
-    variables: Vec<OptAbleVar>,
-}
-
-impl Aliases {
-    pub fn with_capacity(cap: usize) -> Self {
-        Self {
-            aliases: Vec::with_capacity(cap),
-            variables: Vec::with_capacity(cap * 2),
-        }
-    }
-
-    pub fn insert(&mut self, types: &mut Types, symbol: Symbol, alias: Alias) {
-        let alias_variables =
-            {
-                let start = self.variables.len() as _;
-
-                self.variables.extend(
-                    alias
-                        .type_variables
-                        .iter()
-                        .map(|x| OptAbleVar::from(&x.value)),
-                );
-
-                self.variables.extend(alias.lambda_set_variables.iter().map(
-                    |x| match x.as_inner() {
-                        Type::Variable(v) => OptAbleVar::unbound(*v),
-                        _ => unreachable!("lambda set type is not a variable"),
-                    },
-                ));
-
-                let recursion_variables_len = alias.recursion_variables.len() as _;
-                self.variables.extend(
-                    alias
-                        .recursion_variables
-                        .iter()
-                        .copied()
-                        .map(OptAbleVar::unbound),
-                );
-
-                self.variables.extend(
-                    alias
-                        .infer_ext_in_output_variables
-                        .iter()
-                        .map(|v| OptAbleVar::unbound(*v)),
-                );
-
-                DelayedAliasVariables {
-                    start,
-                    type_variables_len: alias.type_variables.len() as _,
-                    lambda_set_variables_len: alias.lambda_set_variables.len() as _,
-                    recursion_variables_len,
-                    infer_ext_in_output_variables_len: alias.infer_ext_in_output_variables.len()
-                        as _,
-                }
-            };
-
-        // TODO: can we construct Aliases from TypeTag directly?
-        let alias_typ = types.from_old_type(&alias.typ);
-
-        self.aliases
-            .push((symbol, alias_typ, alias_variables, alias.kind));
-    }
-
-    fn instantiate_result_result(
-        subs: &mut Subs,
-        rank: Rank,
-        pools: &mut Pools,
-        alias_variables: AliasVariables,
-    ) -> Variable {
-        let tag_names_slice = Subs::RESULT_TAG_NAMES;
-
-        let err_slice = SubsSlice::new(alias_variables.variables_start + 1, 1);
-        let ok_slice = SubsSlice::new(alias_variables.variables_start, 1);
-
-        let variable_slices =
-            SubsSlice::extend_new(&mut subs.variable_slices, [err_slice, ok_slice]);
-
-        let union_tags = UnionTags::from_slices(tag_names_slice, variable_slices);
-        let ext_var = TagExt::Any(Variable::EMPTY_TAG_UNION);
-        let flat_type = FlatType::TagUnion(union_tags, ext_var);
-        let content = Content::Structure(flat_type);
-
-        register(subs, rank, pools, content)
-    }
-
-    /// Build an alias of the form `Num range := range`
-    fn build_num_opaque(
-        subs: &mut Subs,
-        rank: Rank,
-        pools: &mut Pools,
-        symbol: Symbol,
-        range_var: Variable,
-    ) -> Variable {
-        let content = Content::Alias(
-            symbol,
-            AliasVariables::insert_into_subs(subs, [range_var], [], []),
-            range_var,
-            AliasKind::Opaque,
-        );
-
-        register(subs, rank, pools, content)
-    }
-
-    fn instantiate_builtin_aliases_real_var(
-        &mut self,
-        subs: &mut Subs,
-        rank: Rank,
-        pools: &mut Pools,
-        symbol: Symbol,
-        alias_variables: AliasVariables,
-    ) -> Option<(Variable, AliasKind)> {
-        match symbol {
-            Symbol::RESULT_RESULT => {
-                let var = Self::instantiate_result_result(subs, rank, pools, alias_variables);
-
-                Some((var, AliasKind::Structural))
-            }
-            Symbol::NUM_NUM | Symbol::NUM_INTEGER | Symbol::NUM_FLOATINGPOINT => {
-                // Num range := range | Integer range := range | FloatingPoint range := range
-                let range_var = subs.variables[alias_variables.variables_start as usize];
-                Some((range_var, AliasKind::Opaque))
-            }
-            Symbol::NUM_INT => {
-                // Int range : Num (Integer range)
-                //
-                // build `Integer range := range`
-                let integer_content_var = Self::build_num_opaque(
-                    subs,
-                    rank,
-                    pools,
-                    Symbol::NUM_INTEGER,
-                    subs.variables[alias_variables.variables_start as usize],
-                );
-
-                // build `Num (Integer range) := Integer range`
-                let num_content_var =
-                    Self::build_num_opaque(subs, rank, pools, Symbol::NUM_NUM, integer_content_var);
-
-                Some((num_content_var, AliasKind::Structural))
-            }
-            Symbol::NUM_FRAC => {
-                // Frac range : Num (FloatingPoint range)
-                //
-                // build `FloatingPoint range := range`
-                let fpoint_content_var = Self::build_num_opaque(
-                    subs,
-                    rank,
-                    pools,
-                    Symbol::NUM_FLOATINGPOINT,
-                    subs.variables[alias_variables.variables_start as usize],
-                );
-
-                // build `Num (FloatingPoint range) := FloatingPoint range`
-                let num_content_var =
-                    Self::build_num_opaque(subs, rank, pools, Symbol::NUM_NUM, fpoint_content_var);
-
-                Some((num_content_var, AliasKind::Structural))
-            }
-            Symbol::NUM_SIGNED8 => Some((Variable::SIGNED8, AliasKind::Opaque)),
-            Symbol::NUM_SIGNED16 => Some((Variable::SIGNED16, AliasKind::Opaque)),
-            Symbol::NUM_SIGNED32 => Some((Variable::SIGNED32, AliasKind::Opaque)),
-            Symbol::NUM_SIGNED64 => Some((Variable::SIGNED64, AliasKind::Opaque)),
-            Symbol::NUM_SIGNED128 => Some((Variable::SIGNED128, AliasKind::Opaque)),
-            Symbol::NUM_UNSIGNED8 => Some((Variable::UNSIGNED8, AliasKind::Opaque)),
-            Symbol::NUM_UNSIGNED16 => Some((Variable::UNSIGNED16, AliasKind::Opaque)),
-            Symbol::NUM_UNSIGNED32 => Some((Variable::UNSIGNED32, AliasKind::Opaque)),
-            Symbol::NUM_UNSIGNED64 => Some((Variable::UNSIGNED64, AliasKind::Opaque)),
-            Symbol::NUM_UNSIGNED128 => Some((Variable::UNSIGNED128, AliasKind::Opaque)),
-            Symbol::NUM_BINARY32 => Some((Variable::BINARY32, AliasKind::Opaque)),
-            Symbol::NUM_BINARY64 => Some((Variable::BINARY64, AliasKind::Opaque)),
-            _ => None,
-        }
-    }
-
-    fn instantiate_real_var(
-        &mut self,
-        subs: &mut Subs,
-        rank: Rank,
-        pools: &mut Pools,
-        problems: &mut Vec<TypeError>,
-        abilities_store: &AbilitiesStore,
-        obligation_cache: &mut ObligationCache,
-        arena: &bumpalo::Bump,
-        types: &mut Types,
-        symbol: Symbol,
-        alias_variables: AliasVariables,
-    ) -> (Variable, AliasKind) {
-        // hardcoded instantiations for builtin aliases
-        if let Some((var, kind)) = Self::instantiate_builtin_aliases_real_var(
-            self,
-            subs,
-            rank,
-            pools,
-            symbol,
-            alias_variables,
-        ) {
-            return (var, kind);
-        }
-
-        let (typ, delayed_variables, kind) =
-            match self.aliases.iter().find(|(s, _, _, _)| *s == symbol) {
-                None => internal_error!(
-                    "Alias {:?} not registered in delayed aliases! {:?}",
-                    symbol,
-                    &self.aliases
-                ),
-                Some(&(_, typ, delayed_variables, kind)) => (typ, delayed_variables, kind),
-            };
-
-        let mut substitutions: MutMap<_, _> = Default::default();
-
-        let old_type_variables = delayed_variables.type_variables(&mut self.variables);
-        let new_type_variables = &subs.variables[alias_variables.type_variables().indices()];
-
-        for (old, new) in old_type_variables.iter_mut().zip(new_type_variables) {
-            // if constraint gen duplicated a type these variables could be the same
-            // (happens very often in practice)
-            if old.var != *new {
-                substitutions.insert(old.var, *new);
-            }
-        }
-
-        for OptAbleVar {
-            var: rec_var,
-            opt_abilities,
-        } in delayed_variables
-            .recursion_variables(&mut self.variables)
-            .iter_mut()
-        {
-            debug_assert!(opt_abilities.is_none());
-            let new_var = subs.fresh_unnamed_flex_var();
-            substitutions.insert(*rec_var, new_var);
-        }
-
-        let old_lambda_set_variables = delayed_variables.lambda_set_variables(&mut self.variables);
-        let new_lambda_set_variables =
-            &subs.variables[alias_variables.lambda_set_variables().indices()];
-
-        for (old, new) in old_lambda_set_variables
-            .iter_mut()
-            .zip(new_lambda_set_variables)
-        {
-            debug_assert!(old.opt_abilities.is_none());
-            if old.var != *new {
-                substitutions.insert(old.var, *new);
-            }
-        }
-
-        let old_infer_ext_vars =
-            delayed_variables.infer_ext_in_output_variables(&mut self.variables);
-        let new_infer_ext_vars =
-            &subs.variables[alias_variables.infer_ext_in_output_variables().indices()];
-
-        for (old, new) in old_infer_ext_vars.iter_mut().zip(new_infer_ext_vars) {
-            debug_assert!(old.opt_abilities.is_none());
-            if old.var != *new {
-                substitutions.insert(old.var, *new);
-            }
-        }
-
-        let typ = if !substitutions.is_empty() {
-            types.clone_with_variable_substitutions(typ, &substitutions)
-        } else {
-            typ
-        };
-
-        let alias_variable = type_to_variable(
-            subs,
-            rank,
-            pools,
-            problems,
-            abilities_store,
-            obligation_cache,
-            arena,
-            self,
-            types,
-            typ,
-            false,
-        );
-        (alias_variable, kind)
-    }
-}
-
-const DEFAULT_POOLS: usize = 8;
-
-#[derive(Clone, Debug)]
-pub struct Pools(Vec<Vec<Variable>>);
-
-impl Default for Pools {
-    fn default() -> Self {
-        Pools::new(DEFAULT_POOLS)
-    }
-}
-
-impl Pools {
-    pub fn new(num_pools: usize) -> Self {
-        Pools(vec![Vec::new(); num_pools])
-    }
-
-    pub fn len(&self) -> usize {
-        self.0.len()
-    }
-
-    pub fn is_empty(&self) -> bool {
-        self.0.is_empty()
-    }
-
-    pub fn get_mut(&mut self, rank: Rank) -> &mut Vec<Variable> {
-        match self.0.get_mut(rank.into_usize()) {
-            Some(reference) => reference,
-            None => panic!("Compiler bug: could not find pool at rank {}", rank),
-        }
-    }
-
-    pub fn get(&self, rank: Rank) -> &Vec<Variable> {
-        match self.0.get(rank.into_usize()) {
-            Some(reference) => reference,
-            None => panic!("Compiler bug: could not find pool at rank {}", rank),
-        }
-    }
-
-    pub fn iter(&self) -> std::slice::Iter<'_, Vec<Variable>> {
-        self.0.iter()
-    }
-
-    pub fn split_last(mut self) -> (Vec<Variable>, Vec<Vec<Variable>>) {
-        let last = self
-            .0
-            .pop()
-            .unwrap_or_else(|| panic!("Attempted to split_last() on non-empty Pools"));
-
-        (last, self.0)
-    }
-
-    pub fn extend_to(&mut self, n: usize) {
-        for _ in self.len()..n {
-            self.0.push(Vec::new());
-        }
-    }
-}
 
 #[derive(Clone)]
 struct State {
@@ -2264,1294 +1871,6 @@ impl LocalDefVarsVec<(Symbol, Loc<Variable>)> {
     }
 }
 
-use std::cell::RefCell;
-use std::ops::ControlFlow;
-
-std::thread_local! {
-    /// Scratchpad arena so we don't need to allocate a new one all the time
-    static SCRATCHPAD: RefCell<Option<bumpalo::Bump>> = RefCell::new(Some(bumpalo::Bump::with_capacity(4 * 1024)));
-}
-
-fn take_scratchpad() -> bumpalo::Bump {
-    SCRATCHPAD.with(|f| f.take().unwrap())
-}
-
-fn put_scratchpad(scratchpad: bumpalo::Bump) {
-    SCRATCHPAD.with(|f| {
-        f.replace(Some(scratchpad));
-    });
-}
-
-fn either_type_index_to_var(
-    subs: &mut Subs,
-    rank: Rank,
-    pools: &mut Pools,
-    problems: &mut Vec<TypeError>,
-    abilities_store: &mut AbilitiesStore,
-    obligation_cache: &mut ObligationCache,
-    types: &mut Types,
-    aliases: &mut Aliases,
-    either_type_index: TypeOrVar,
-) -> Variable {
-    match either_type_index.split() {
-        Ok(type_index) => {
-            // Converts the celled type to a variable, emplacing the new variable for re-use.
-            let var = type_to_var(
-                subs,
-                rank,
-                problems,
-                abilities_store,
-                obligation_cache,
-                pools,
-                types,
-                aliases,
-                type_index,
-            );
-
-            debug_assert!(
-                matches!(types[type_index], TypeTag::Variable(v) if v == var)
-                    || matches!(
-                        types[type_index],
-                        TypeTag::EmptyRecord | TypeTag::EmptyTagUnion
-                    )
-            );
-            var
-        }
-        Err(var_index) => {
-            // we cheat, and  store the variable directly in the index
-            unsafe { Variable::from_index(var_index.index() as _) }
-        }
-    }
-}
-
-pub(crate) fn type_to_var(
-    subs: &mut Subs,
-    rank: Rank,
-    problems: &mut Vec<TypeError>,
-    abilities_store: &mut AbilitiesStore,
-    obligation_cache: &mut ObligationCache,
-    pools: &mut Pools,
-    types: &mut Types,
-    aliases: &mut Aliases,
-    typ: Index<TypeTag>,
-) -> Variable {
-    if let TypeTag::Variable(var) = types[typ] {
-        var
-    } else {
-        let mut arena = take_scratchpad();
-
-        let var = type_to_variable(
-            subs,
-            rank,
-            pools,
-            problems,
-            abilities_store,
-            obligation_cache,
-            &arena,
-            aliases,
-            types,
-            typ,
-            false,
-        );
-
-        arena.reset();
-        put_scratchpad(arena);
-
-        var
-    }
-}
-
-enum RegisterVariable {
-    /// Based on the Type, we already know what variable this will be
-    Direct(Variable),
-    /// This Type needs more complicated Content. We reserve a Variable
-    /// for it, but put a placeholder Content in subs
-    Deferred,
-}
-
-impl RegisterVariable {
-    fn from_type(
-        subs: &mut Subs,
-        rank: Rank,
-        pools: &mut Pools,
-        arena: &'_ bumpalo::Bump,
-        types: &mut Types,
-        typ: Index<TypeTag>,
-    ) -> Self {
-        use RegisterVariable::*;
-
-        match types[typ] {
-            TypeTag::Variable(var) => Direct(var),
-            TypeTag::EmptyRecord => Direct(Variable::EMPTY_RECORD),
-            TypeTag::EmptyTagUnion => Direct(Variable::EMPTY_TAG_UNION),
-            TypeTag::DelayedAlias { shared }
-            | TypeTag::StructuralAlias { shared, .. }
-            | TypeTag::OpaqueAlias { shared, .. }
-            | TypeTag::HostExposedAlias { shared, .. } => {
-                let AliasShared { symbol, .. } = types[shared];
-                if let Some(reserved) = Variable::get_reserved(symbol) {
-                    let direct_var = if rank.is_generalized() {
-                        // reserved variables are stored with rank NONE
-                        reserved
-                    } else {
-                        // for any other rank, we need to copy; it takes care of adjusting the rank
-                        deep_copy_var_in(subs, rank, pools, reserved, arena)
-                    };
-                    // Safety: the `destination` will become the source-of-truth for the type index, since it
-                    // was not already transformed before (if it was, we'd be in the Variable branch!)
-                    let _old_typ = unsafe { types.emplace_variable(typ, direct_var) };
-                    return Direct(direct_var);
-                }
-
-                Deferred
-            }
-            _ => Deferred,
-        }
-    }
-
-    #[inline(always)]
-    fn with_stack(
-        subs: &mut Subs,
-        rank: Rank,
-        pools: &mut Pools,
-        arena: &'_ bumpalo::Bump,
-        types: &mut Types,
-        typ_index: Index<TypeTag>,
-        stack: &mut bumpalo::collections::Vec<'_, TypeToVar>,
-    ) -> Variable {
-        match Self::from_type(subs, rank, pools, arena, types, typ_index) {
-            Self::Direct(var) => var,
-            Self::Deferred => {
-                let var = subs.fresh_unnamed_flex_var();
-                // Safety: the `destination` will become the source-of-truth for the type index, since it
-                // was not already transformed before (if it was, it wouldn't be deferred!)
-                let typ = unsafe { types.emplace_variable(typ_index, var) };
-                stack.push(TypeToVar::Defer {
-                    typ,
-                    typ_index,
-                    destination: var,
-                    ambient_function: AmbientFunctionPolicy::NoFunction,
-                });
-                var
-            }
-        }
-    }
-}
-
-/// Instantiation of ambient functions in unspecialized lambda sets is somewhat tricky due to other
-/// optimizations we have in place. This struct tells us how they should be instantiated.
-#[derive(Debug)]
-enum AmbientFunctionPolicy {
-    /// We're not in a function. This variant may never hold for unspecialized lambda sets.
-    NoFunction,
-    /// We're in a known function.
-    Function(Variable),
-}
-
-impl AmbientFunctionPolicy {
-    fn link_to_alias_lambda_set_var(&self, subs: &mut Subs, var: Variable) {
-        let ambient_function = match self {
-            AmbientFunctionPolicy::Function(var) => *var,
-            _ => {
-                // Might be linked at a deeper point in time, ignore for now
-                return;
-            }
-        };
-        let content = subs.get_content_without_compacting(var);
-        let new_content = match content {
-            Content::LambdaSet(LambdaSet {
-                solved,
-                recursion_var,
-                unspecialized,
-                ambient_function: _,
-            }) => Content::LambdaSet(LambdaSet {
-                solved: *solved,
-                recursion_var: *recursion_var,
-                unspecialized: *unspecialized,
-                ambient_function,
-            }),
-            Content::FlexVar(_) => {
-                // Something like
-                //   Encoder fmt <a> : List U8, fmt -a-> List U8 | fmt has EncoderFormatting
-                // THEORY: Replace these with empty lambda sets. They will unify the same as a flex
-                // var does, but allows us to record the ambient function properly.
-                Content::LambdaSet(LambdaSet {
-                    solved: UnionLabels::default(),
-                    recursion_var: OptVariable::NONE,
-                    unspecialized: SubsSlice::default(),
-                    ambient_function,
-                })
-            }
-            content => internal_error!("{:?}({:?}) not a lambda set", content, var),
-        };
-        subs.set_content_unchecked(var, new_content);
-    }
-}
-
-#[derive(Debug)]
-enum TypeToVar {
-    Defer {
-        typ: TypeTag,
-        typ_index: Index<TypeTag>,
-        destination: Variable,
-        ambient_function: AmbientFunctionPolicy,
-    },
-}
-
-#[allow(clippy::too_many_arguments)]
-fn type_to_variable(
-    subs: &mut Subs,
-    rank: Rank,
-    pools: &mut Pools,
-    problems: &mut Vec<TypeError>,
-    abilities_store: &AbilitiesStore,
-    obligation_cache: &mut ObligationCache,
-    arena: &bumpalo::Bump,
-    aliases: &mut Aliases,
-    types: &mut Types,
-    typ: Index<TypeTag>,
-    // Helpers for instantiating ambient functions of lambda set variables from type aliases.
-    is_alias_lambda_set_arg: bool,
-) -> Variable {
-    use bumpalo::collections::Vec;
-
-    let mut stack = Vec::with_capacity_in(8, arena);
-    let mut bind_to_abilities = Vec::new_in(arena);
-
-    macro_rules! helper {
-        ($typ:expr, $ambient_function_policy:expr) => {{
-            match RegisterVariable::from_type(subs, rank, pools, arena, types, $typ) {
-                RegisterVariable::Direct(var) => {
-                    // If the variable is just a type variable but we know we're in a lambda set
-                    // context, try to link to the ambient function.
-                    $ambient_function_policy.link_to_alias_lambda_set_var(subs, var);
-
-                    var
-                }
-                RegisterVariable::Deferred => {
-                    let var = subs.fresh_unnamed_flex_var();
-
-                    // Safety: the `destination` will become the source-of-truth for the type index, since it
-                    // was not already transformed before (if it was, it wouldn't be deferred!)
-                    let typ = unsafe { types.emplace_variable($typ, var) };
-
-                    stack.push(TypeToVar::Defer {
-                        typ,
-                        typ_index: $typ,
-                        destination: var,
-                        ambient_function: $ambient_function_policy,
-                    });
-
-                    var
-                }
-            }
-        }};
-        ($typ:expr) => {{
-            helper!($typ, AmbientFunctionPolicy::NoFunction)
-        }};
-    }
-
-    let result = helper!(typ);
-
-    while let Some(TypeToVar::Defer {
-        typ_index,
-        typ,
-        destination,
-        ambient_function,
-    }) = stack.pop()
-    {
-        use TypeTag::*;
-        match typ {
-            Variable(_) | EmptyRecord | EmptyTagUnion => {
-                unreachable!("This variant should never be deferred!",)
-            }
-            RangedNumber(range) => {
-                let content = Content::RangedNumber(range);
-
-                register_with_known_var(subs, destination, rank, pools, content)
-            }
-            Apply {
-                symbol,
-                type_argument_regions: _,
-                region: _,
-            } => {
-                let arguments = types.get_type_arguments(typ_index);
-                let new_arguments = VariableSubsSlice::reserve_into_subs(subs, arguments.len());
-                for (target_index, var_index) in
-                    (new_arguments.indices()).zip(arguments.into_iter())
-                {
-                    let var = helper!(var_index);
-                    subs.variables[target_index] = var;
-                }
-
-                let flat_type = FlatType::Apply(symbol, new_arguments);
-                let content = Content::Structure(flat_type);
-
-                register_with_known_var(subs, destination, rank, pools, content)
-            }
-
-            ClosureTag {
-                name,
-                ambient_function,
-            } => {
-                let captures = types.get_type_arguments(typ_index);
-                let union_lambdas = create_union_lambda(
-                    subs, rank, pools, arena, types, name, captures, &mut stack,
-                );
-
-                let content = Content::LambdaSet(subs::LambdaSet {
-                    solved: union_lambdas,
-                    // We may figure out the lambda set is recursive during solving, but it never
-                    // is to begin with.
-                    recursion_var: OptVariable::NONE,
-                    unspecialized: SubsSlice::default(),
-                    ambient_function,
-                });
-
-                register_with_known_var(subs, destination, rank, pools, content)
-            }
-            UnspecializedLambdaSet { unspecialized } => {
-                let unspecialized_slice = SubsSlice::extend_new(
-                    &mut subs.unspecialized_lambda_sets,
-                    std::iter::once(unspecialized),
-                );
-
-                // `ClosureTag` ambient functions are resolved during constraint generation.
-                // But `UnspecializedLambdaSet`s can only ever live in a type signature, and don't
-                // correspond to a expression, so they are never constrained.
-                // Instead, we resolve their ambient functions during type translation, observing
-                // the invariant that a lambda set can only ever appear under a function type.
-                let ambient_function = match ambient_function {
-                    AmbientFunctionPolicy::NoFunction => {
-                        debug_assert!(is_alias_lambda_set_arg);
-                        // To be filled in during delayed type alias instantiation
-                        roc_types::subs::Variable::NULL
-                    }
-                    AmbientFunctionPolicy::Function(var) => var,
-                };
-
-                let content = Content::LambdaSet(subs::LambdaSet {
-                    unspecialized: unspecialized_slice,
-                    solved: UnionLabels::default(),
-                    recursion_var: OptVariable::NONE,
-                    ambient_function,
-                });
-
-                register_with_known_var(subs, destination, rank, pools, content)
-            }
-            // This case is important for the rank of boolean variables
-            Function(closure_type, ret_type) => {
-                let arguments = types.get_type_arguments(typ_index);
-                let new_arguments = VariableSubsSlice::reserve_into_subs(subs, arguments.len());
-                for (target_index, var_index) in
-                    (new_arguments.indices()).zip(arguments.into_iter())
-                {
-                    let var = helper!(var_index);
-                    subs.variables[target_index] = var;
-                }
-
-                let ret_var = helper!(ret_type);
-                let closure_var =
-                    helper!(closure_type, AmbientFunctionPolicy::Function(destination));
-                let content =
-                    Content::Structure(FlatType::Func(new_arguments, closure_var, ret_var));
-
-                register_with_known_var(subs, destination, rank, pools, content)
-            }
-            Record(fields) => {
-                let ext_slice = types.get_type_arguments(typ_index);
-
-                // An empty fields is inefficient (but would be correct)
-                // If hit, try to turn the value into an EmptyRecord in canonicalization
-                debug_assert!(!fields.is_empty() || !ext_slice.is_empty());
-
-                let mut field_vars = Vec::with_capacity_in(fields.len(), arena);
-
-                let (fields_names, field_kinds, field_tys) = types.record_fields_slices(fields);
-
-                for ((field, field_kind), field_type) in (fields_names.into_iter())
-                    .zip(field_kinds.into_iter())
-                    .zip(field_tys.into_iter())
-                {
-                    let field_var = {
-                        let t = helper!(field_type);
-                        types[field_kind].replace(t)
-                    };
-
-                    field_vars.push((types[field].clone(), field_var));
-                }
-
-                debug_assert!(ext_slice.len() <= 1);
-                let temp_ext_var = match ext_slice.into_iter().next() {
-                    None => roc_types::subs::Variable::EMPTY_RECORD,
-                    Some(ext) => helper!(ext),
-                };
-
-                let (it, new_ext_var) =
-                    gather_fields_unsorted_iter(subs, RecordFields::empty(), temp_ext_var)
-                        .expect("Something ended up weird in this record type");
-
-                let it = it
-                    .into_iter()
-                    .map(|(field, field_type)| (field.clone(), field_type));
-
-                field_vars.extend(it);
-                insertion_sort_by(&mut field_vars, RecordFields::compare);
-
-                let record_fields = RecordFields::insert_into_subs(subs, field_vars);
-
-                let content = Content::Structure(FlatType::Record(record_fields, new_ext_var));
-
-                register_with_known_var(subs, destination, rank, pools, content)
-            }
-
-            Tuple(elems) => {
-                let ext_slice = types.get_type_arguments(typ_index);
-
-                // Elems should never be empty; we don't support empty tuples
-                debug_assert!(!elems.is_empty() || !ext_slice.is_empty());
-
-                let mut elem_vars = Vec::with_capacity_in(elems.len(), arena);
-
-                let (indices, elem_tys) = types.tuple_elems_slices(elems);
-
-                for (index, elem_type) in indices.into_iter().zip(elem_tys.into_iter()) {
-                    let elem_var = helper!(elem_type);
-                    elem_vars.push((types[index], elem_var));
-                }
-
-                debug_assert!(ext_slice.len() <= 1);
-                let temp_ext_var = match ext_slice.into_iter().next() {
-                    None => roc_types::subs::Variable::EMPTY_TUPLE,
-                    Some(ext) => helper!(ext),
-                };
-
-                let (it, new_ext_var) =
-                    gather_tuple_elems_unsorted_iter(subs, TupleElems::empty(), temp_ext_var)
-                        .expect("Something ended up weird in this tuple type");
-
-                elem_vars.extend(it);
-                let tuple_elems = TupleElems::insert_into_subs(subs, elem_vars);
-
-                let content = Content::Structure(FlatType::Tuple(tuple_elems, new_ext_var));
-
-                register_with_known_var(subs, destination, rank, pools, content)
-            }
-
-            TagUnion(tags, ext_openness) => {
-                let ext_slice = types.get_type_arguments(typ_index);
-
-                // An empty tags is inefficient (but would be correct)
-                // If hit, try to turn the value into an EmptyTagUnion in canonicalization
-                debug_assert!(!tags.is_empty() || !ext_slice.is_empty());
-
-                let (union_tags, ext) = type_to_union_tags(
-                    subs,
-                    rank,
-                    pools,
-                    arena,
-                    types,
-                    tags,
-                    ext_slice,
-                    ext_openness,
-                    &mut stack,
-                );
-                let content = Content::Structure(FlatType::TagUnion(union_tags, ext));
-
-                register_with_known_var(subs, destination, rank, pools, content)
-            }
-            FunctionOrTagUnion(symbol, ext_openness) => {
-                let ext_slice = types.get_type_arguments(typ_index);
-                let tag_name = types.get_tag_name(&typ_index).clone();
-
-                debug_assert!(ext_slice.len() <= 1);
-                let temp_ext = match ext_slice.into_iter().next() {
-                    Some(ext) => {
-                        let var = helper!(ext);
-                        TagExt::from_can(var, ext_openness)
-                    }
-                    None => TagExt::Any(roc_types::subs::Variable::EMPTY_TAG_UNION),
-                };
-
-                let (it, ext) = roc_types::types::gather_tags_unsorted_iter(
-                    subs,
-                    UnionTags::default(),
-                    temp_ext,
-                )
-                .expect("extension var could not be seen as a tag union");
-
-                for _ in it {
-                    unreachable!("we assert that the ext var is empty; otherwise we'd already know it was a tag union!");
-                }
-
-                let tag_names = SubsSlice::extend_new(&mut subs.tag_names, [tag_name]);
-                let symbols = SubsSlice::extend_new(&mut subs.symbol_names, [symbol]);
-
-                let content =
-                    Content::Structure(FlatType::FunctionOrTagUnion(tag_names, symbols, ext));
-
-                register_with_known_var(subs, destination, rank, pools, content)
-            }
-            RecursiveTagUnion(rec_var, tags, ext_openness) => {
-                let ext_slice = types.get_type_arguments(typ_index);
-
-                // An empty tags is inefficient (but would be correct)
-                // If hit, try to turn the value into an EmptyTagUnion in canonicalization
-                debug_assert!(!tags.is_empty() || !ext_slice.is_empty());
-
-                let (union_tags, ext) = type_to_union_tags(
-                    subs,
-                    rank,
-                    pools,
-                    arena,
-                    types,
-                    tags,
-                    ext_slice,
-                    ext_openness,
-                    &mut stack,
-                );
-                let content =
-                    Content::Structure(FlatType::RecursiveTagUnion(rec_var, union_tags, ext));
-
-                let tag_union_var = destination;
-                register_with_known_var(subs, tag_union_var, rank, pools, content);
-
-                register_with_known_var(
-                    subs,
-                    rec_var,
-                    rank,
-                    pools,
-                    Content::RecursionVar {
-                        opt_name: None,
-                        structure: tag_union_var,
-                    },
-                );
-
-                tag_union_var
-            }
-
-            DelayedAlias { shared } => {
-                let AliasShared {
-                    symbol,
-                    type_argument_abilities,
-                    type_argument_regions,
-                    lambda_set_variables,
-                    infer_ext_in_output_variables,
-                } = types[shared];
-
-                let type_arguments = types.get_type_arguments(typ_index);
-
-                let alias_variables = {
-                    let all_vars_length = type_arguments.len()
-                        + lambda_set_variables.len()
-                        + infer_ext_in_output_variables.len();
-                    let new_variables = VariableSubsSlice::reserve_into_subs(subs, all_vars_length);
-
-                    let type_arguments_offset = 0;
-                    let lambda_set_vars_offset = type_arguments_offset + type_arguments.len();
-                    let infer_ext_vars_offset = lambda_set_vars_offset + lambda_set_variables.len();
-
-                    for (((target_index, arg_type), arg_region), abilities) in
-                        (new_variables.indices().skip(type_arguments_offset))
-                            .zip(type_arguments.into_iter())
-                            .zip(type_argument_regions.into_iter())
-                            .zip(type_argument_abilities.into_iter())
-                    {
-                        let copy_var = helper!(arg_type);
-                        subs.variables[target_index] = copy_var;
-                        if !types[abilities].is_empty() {
-                            let arg_region = types[arg_region];
-                            bind_to_abilities.push((Loc::at(arg_region, copy_var), abilities));
-                        }
-                    }
-
-                    let it = (new_variables.indices().skip(lambda_set_vars_offset))
-                        .zip(lambda_set_variables.into_iter());
-                    for (target_index, ls) in it {
-                        // We MUST do this now, otherwise when linking the ambient function during
-                        // instantiation of the real var, there will be nothing to link against.
-                        let copy_var = type_to_variable(
-                            subs,
-                            rank,
-                            pools,
-                            problems,
-                            abilities_store,
-                            obligation_cache,
-                            arena,
-                            aliases,
-                            types,
-                            ls,
-                            true,
-                        );
-                        subs.variables[target_index] = copy_var;
-                    }
-
-                    let it = (new_variables.indices().skip(infer_ext_vars_offset))
-                        .zip(infer_ext_in_output_variables.into_iter());
-                    for (target_index, ext_typ) in it {
-                        let copy_var = helper!(ext_typ);
-                        subs.variables[target_index] = copy_var;
-                    }
-
-                    AliasVariables {
-                        variables_start: new_variables.start,
-                        type_variables_len: type_arguments.len() as _,
-                        lambda_set_variables_len: lambda_set_variables.len() as _,
-                        all_variables_len: all_vars_length as _,
-                    }
-                };
-
-                let (alias_variable, kind) = aliases.instantiate_real_var(
-                    subs,
-                    rank,
-                    pools,
-                    problems,
-                    abilities_store,
-                    obligation_cache,
-                    arena,
-                    types,
-                    symbol,
-                    alias_variables,
-                );
-
-                let content = Content::Alias(symbol, alias_variables, alias_variable, kind);
-
-                register_with_known_var(subs, destination, rank, pools, content)
-            }
-
-            StructuralAlias { shared, actual } | OpaqueAlias { shared, actual } => {
-                let kind = match typ {
-                    StructuralAlias { .. } => AliasKind::Structural,
-                    OpaqueAlias { .. } => AliasKind::Opaque,
-                    _ => internal_error!(),
-                };
-
-                let AliasShared {
-                    symbol,
-                    type_argument_abilities,
-                    type_argument_regions,
-                    lambda_set_variables,
-                    infer_ext_in_output_variables,
-                } = types[shared];
-
-                debug_assert!(roc_types::subs::Variable::get_reserved(symbol).is_none());
-
-                let type_arguments = types.get_type_arguments(typ_index);
-
-                let alias_variables = {
-                    let all_vars_length = type_arguments.len()
-                        + lambda_set_variables.len()
-                        + infer_ext_in_output_variables.len();
-
-                    let type_arguments_offset = 0;
-                    let lambda_set_vars_offset = type_arguments_offset + type_arguments.len();
-                    let infer_ext_vars_offset = lambda_set_vars_offset + lambda_set_variables.len();
-
-                    let new_variables = VariableSubsSlice::reserve_into_subs(subs, all_vars_length);
-
-                    for (((target_index, typ), region), abilities) in
-                        (new_variables.indices().skip(type_arguments_offset))
-                            .zip(type_arguments.into_iter())
-                            .zip(type_argument_regions.into_iter())
-                            .zip(type_argument_abilities.into_iter())
-                    {
-                        let copy_var = helper!(typ);
-                        subs.variables[target_index] = copy_var;
-                        if !types[abilities].is_empty() {
-                            let region = types[region];
-                            bind_to_abilities.push((Loc::at(region, copy_var), abilities));
-                        }
-                    }
-
-                    let it = (new_variables.indices().skip(lambda_set_vars_offset))
-                        .zip(lambda_set_variables.into_iter());
-                    for (target_index, ls) in it {
-                        let copy_var = helper!(ls);
-                        subs.variables[target_index] = copy_var;
-                    }
-
-                    let it = (new_variables.indices().skip(infer_ext_vars_offset))
-                        .zip(infer_ext_in_output_variables.into_iter());
-                    for (target_index, ext_typ) in it {
-                        let copy_var = helper!(ext_typ);
-                        subs.variables[target_index] = copy_var;
-                    }
-
-                    AliasVariables {
-                        variables_start: new_variables.start,
-                        type_variables_len: type_arguments.len() as _,
-                        lambda_set_variables_len: lambda_set_variables.len() as _,
-                        all_variables_len: all_vars_length as _,
-                    }
-                };
-
-                let alias_variable = if let Symbol::RESULT_RESULT = symbol {
-                    roc_result_to_var(subs, rank, pools, arena, types, actual, &mut stack)
-                } else {
-                    helper!(actual)
-                };
-                let content = Content::Alias(symbol, alias_variables, alias_variable, kind);
-
-                register_with_known_var(subs, destination, rank, pools, content)
-            }
-            HostExposedAlias {
-                shared,
-                actual_type: alias_type,
-                actual_variable: actual_var,
-            } => {
-                let AliasShared {
-                    symbol,
-                    type_argument_abilities: _,
-                    type_argument_regions: _,
-                    lambda_set_variables,
-                    infer_ext_in_output_variables: _, // TODO
-                } = types[shared];
-
-                let type_arguments = types.get_type_arguments(typ_index);
-
-                let alias_variables = {
-                    let length = type_arguments.len() + lambda_set_variables.len();
-                    let new_variables = VariableSubsSlice::reserve_into_subs(subs, length);
-
-                    for (target_index, arg_type) in
-                        (new_variables.indices()).zip(type_arguments.into_iter())
-                    {
-                        let copy_var = helper!(arg_type);
-                        subs.variables[target_index] = copy_var;
-                    }
-                    let it = (new_variables.indices().skip(type_arguments.len()))
-                        .zip(lambda_set_variables.into_iter());
-                    for (target_index, ls) in it {
-                        // We MUST do this now, otherwise when linking the ambient function during
-                        // instantiation of the real var, there will be nothing to link against.
-                        let copy_var = type_to_variable(
-                            subs,
-                            rank,
-                            pools,
-                            problems,
-                            abilities_store,
-                            obligation_cache,
-                            arena,
-                            aliases,
-                            types,
-                            ls,
-                            true,
-                        );
-                        subs.variables[target_index] = copy_var;
-                    }
-
-                    AliasVariables {
-                        variables_start: new_variables.start,
-                        type_variables_len: type_arguments.len() as _,
-                        lambda_set_variables_len: lambda_set_variables.len() as _,
-                        all_variables_len: length as _,
-                    }
-                };
-
-                // cannot use helper! here because this variable may be involved in unification below
-                let alias_variable = type_to_variable(
-                    subs,
-                    rank,
-                    pools,
-                    problems,
-                    abilities_store,
-                    obligation_cache,
-                    arena,
-                    aliases,
-                    types,
-                    alias_type,
-                    false,
-                );
-                // TODO(opaques): I think host-exposed aliases should always be structural
-                // (when does it make sense to give a host an opaque type?)
-                let content = Content::Alias(
-                    symbol,
-                    alias_variables,
-                    alias_variable,
-                    AliasKind::Structural,
-                );
-                let result = register_with_known_var(subs, destination, rank, pools, content);
-
-                // We only want to unify the actual_var with the alias once
-                // if it's already redirected (and therefore, redundant)
-                // don't do it again
-                if !subs.redundant(actual_var) {
-                    let descriptor = subs.get(result);
-                    subs.union(result, actual_var, descriptor);
-                }
-
-                result
-            }
-            Error => {
-                let content = Content::Error;
-
-                register_with_known_var(subs, destination, rank, pools, content)
-            }
-        };
-    }
-
-    for (Loc { value: var, region }, abilities) in bind_to_abilities {
-        let abilities = &types[abilities];
-        match *subs.get_content_unchecked(var) {
-            Content::RigidVar(a) => {
-                // TODO(multi-abilities): check run cache
-                let abilities_slice =
-                    SubsSlice::extend_new(&mut subs.symbol_names, abilities.sorted_iter().copied());
-                subs.set_content(var, Content::RigidAbleVar(a, abilities_slice));
-            }
-            Content::RigidAbleVar(_, abs)
-                if (subs.get_subs_slice(abs).iter()).eq(abilities.sorted_iter()) =>
-            {
-                // pass, already bound
-            }
-            _ => {
-                let abilities_slice =
-                    SubsSlice::extend_new(&mut subs.symbol_names, abilities.sorted_iter().copied());
-
-                let flex_ability = register(
-                    subs,
-                    rank,
-                    pools,
-                    Content::FlexAbleVar(None, abilities_slice),
-                );
-
-                let category = Category::OpaqueArg;
-                match unify(
-                    &mut UEnv::new(subs),
-                    var,
-                    flex_ability,
-                    Mode::EQ,
-                    Polarity::OF_VALUE,
-                ) {
-                    Success {
-                        vars: _,
-                        must_implement_ability,
-                        lambda_sets_to_specialize,
-                        extra_metadata: _,
-                    } => {
-                        // No introduction needed
-
-                        if !must_implement_ability.is_empty() {
-                            let new_problems = obligation_cache.check_obligations(
-                                subs,
-                                abilities_store,
-                                must_implement_ability,
-                                AbilityImplError::BadExpr(region, category, flex_ability),
-                            );
-                            problems.extend(new_problems);
-                        }
-                        debug_assert!(lambda_sets_to_specialize
-                            .drain()
-                            .all(|(_, vals)| vals.is_empty()));
-                    }
-                    Failure(_vars, actual_type, expected_type, _bad_impls) => {
-                        // No introduction needed
-
-                        let problem = TypeError::BadExpr(
-                            region,
-                            category,
-                            actual_type,
-                            Expected::NoExpectation(expected_type),
-                        );
-
-                        problems.push(problem);
-                    }
-                }
-            }
-        }
-    }
-
-    result
-}
-
-#[inline(always)]
-fn roc_result_to_var(
-    subs: &mut Subs,
-    rank: Rank,
-    pools: &mut Pools,
-    arena: &'_ bumpalo::Bump,
-    types: &mut Types,
-    result_type: Index<TypeTag>,
-    stack: &mut bumpalo::collections::Vec<'_, TypeToVar>,
-) -> Variable {
-    match types[result_type] {
-        TypeTag::TagUnion(tags, _ext_openness) => {
-            let ext_slice = types.get_type_arguments(result_type);
-
-            debug_assert!(ext_slice.is_empty());
-            debug_assert!(tags.len() == 2);
-
-            let (tags_slice, payload_slices_slice) = types.union_tag_slices(tags);
-
-            if let ([err, ok], [err_args, ok_args]) =
-                (&types[tags_slice], &types[payload_slices_slice])
-            {
-                debug_assert_eq!(err, &subs.tag_names[0]);
-                debug_assert_eq!(ok, &subs.tag_names[1]);
-
-                debug_assert_eq!(err_args.len(), 1);
-                debug_assert_eq!(ok_args.len(), 1);
-
-                if let (Some(err_type), Some(ok_type)) =
-                    (err_args.into_iter().next(), ok_args.into_iter().next())
-                {
-                    let err_var = RegisterVariable::with_stack(
-                        subs, rank, pools, arena, types, err_type, stack,
-                    );
-                    let ok_var = RegisterVariable::with_stack(
-                        subs, rank, pools, arena, types, ok_type, stack,
-                    );
-
-                    let start = subs.variables.len() as u32;
-                    let err_slice = SubsSlice::new(start, 1);
-                    let ok_slice = SubsSlice::new(start + 1, 1);
-
-                    subs.variables.push(err_var);
-                    subs.variables.push(ok_var);
-
-                    let variables = SubsSlice::new(subs.variable_slices.len() as _, 2);
-                    subs.variable_slices.push(err_slice);
-                    subs.variable_slices.push(ok_slice);
-
-                    let union_tags = UnionTags::from_slices(Subs::RESULT_TAG_NAMES, variables);
-                    let ext = TagExt::Any(Variable::EMPTY_TAG_UNION);
-
-                    let content = Content::Structure(FlatType::TagUnion(union_tags, ext));
-
-                    return register(subs, rank, pools, content);
-                }
-            }
-
-            unreachable!("invalid arguments to Result.Result; canonicalization should catch this!")
-        }
-        _ => unreachable!("not a valid type inside a Result.Result alias"),
-    }
-}
-
-fn insertion_sort_by<T, F>(arr: &mut [T], mut compare: F)
-where
-    F: FnMut(&T, &T) -> std::cmp::Ordering,
-{
-    for i in 1..arr.len() {
-        let val = &arr[i];
-        let mut j = i;
-        let pos = arr[..i]
-            .binary_search_by(|x| compare(x, val))
-            .unwrap_or_else(|pos| pos);
-        // Swap all elements until specific position.
-        while j > pos {
-            arr.swap(j - 1, j);
-            j -= 1;
-        }
-    }
-}
-
-fn sorted_no_duplicate_tags(tag_slices: &[TagName]) -> bool {
-    match tag_slices.split_first() {
-        None => true,
-        Some((first, rest)) => {
-            let mut current = first;
-
-            for next in rest {
-                if current >= next {
-                    return false;
-                } else {
-                    current = next;
-                }
-            }
-
-            true
-        }
-    }
-}
-
-fn sort_and_deduplicate<T>(tag_vars: &mut bumpalo::collections::Vec<(TagName, T)>) {
-    insertion_sort_by(tag_vars, |(a, _), (b, _)| a.cmp(b));
-
-    // deduplicate, keeping the right-most occurrence of a tag name
-    let mut i = 0;
-
-    while i < tag_vars.len() {
-        match (tag_vars.get(i), tag_vars.get(i + 1)) {
-            (Some((t1, _)), Some((t2, _))) => {
-                if t1 == t2 {
-                    tag_vars.remove(i);
-                } else {
-                    i += 1;
-                }
-            }
-            _ => break,
-        }
-    }
-}
-
-/// Find whether the current run of tag names is in the subs.tag_names array already. If so,
-/// we take a SubsSlice to the existing tag names, so we don't have to add/clone those tag names
-/// and keep subs memory consumption low
-fn find_tag_name_run(slice: &[TagName], subs: &mut Subs) -> Option<SubsSlice<TagName>> {
-    use std::cmp::Ordering;
-
-    let tag_name = slice.get(0)?;
-
-    let mut result = None;
-
-    // the `SubsSlice<TagName>` that inserting `slice` into subs would give
-    let bigger_slice = SubsSlice::new(subs.tag_names.len() as _, slice.len() as _);
-
-    match subs.tag_name_cache.get_mut(tag_name) {
-        Some(occupied) => {
-            let subs_slice = *occupied;
-
-            let prefix_slice = SubsSlice::new(subs_slice.start, slice.len() as _);
-
-            if slice.len() == 1 {
-                return Some(prefix_slice);
-            }
-
-            match slice.len().cmp(&subs_slice.len()) {
-                Ordering::Less => {
-                    // we might have a prefix
-                    let tag_names = &subs.tag_names[subs_slice.start as usize..];
-
-                    for (from_subs, from_slice) in tag_names.iter().zip(slice.iter()) {
-                        if from_subs != from_slice {
-                            return None;
-                        }
-                    }
-
-                    result = Some(prefix_slice);
-                }
-                Ordering::Equal => {
-                    let tag_names = &subs.tag_names[subs_slice.indices()];
-
-                    for (from_subs, from_slice) in tag_names.iter().zip(slice.iter()) {
-                        if from_subs != from_slice {
-                            return None;
-                        }
-                    }
-
-                    result = Some(subs_slice);
-                }
-                Ordering::Greater => {
-                    // switch to the bigger slice that is not inserted yet, but will be soon
-                    *occupied = bigger_slice;
-                }
-            }
-        }
-        None => {
-            subs.tag_name_cache.push(tag_name, bigger_slice);
-        }
-    }
-
-    result
-}
-
-#[inline(always)]
-fn register_tag_arguments(
-    subs: &mut Subs,
-    rank: Rank,
-    pools: &mut Pools,
-    arena: &'_ bumpalo::Bump,
-    types: &mut Types,
-    stack: &mut bumpalo::collections::Vec<'_, TypeToVar>,
-    arguments: Slice<TypeTag>,
-) -> VariableSubsSlice {
-    if arguments.is_empty() {
-        VariableSubsSlice::default()
-    } else {
-        let new_variables = VariableSubsSlice::reserve_into_subs(subs, arguments.len());
-        let it = new_variables.indices().zip(arguments.into_iter());
-
-        for (target_index, argument) in it {
-            let var =
-                RegisterVariable::with_stack(subs, rank, pools, arena, types, argument, stack);
-            subs.variables[target_index] = var;
-        }
-
-        new_variables
-    }
-}
-
-/// Assumes that the tags are sorted and there are no duplicates!
-fn insert_tags_fast_path(
-    subs: &mut Subs,
-    rank: Rank,
-    pools: &mut Pools,
-    arena: &'_ bumpalo::Bump,
-    types: &mut Types,
-    union_tags: UnionTags,
-    stack: &mut bumpalo::collections::Vec<'_, TypeToVar>,
-) -> UnionTags {
-    let (tags, payload_slices) = types.union_tag_slices(union_tags);
-
-    debug_assert_eq!(tags.len(), payload_slices.len());
-
-    if let [arguments_slice] = &types[payload_slices] {
-        let arguments_slice = *arguments_slice;
-
-        let variable_slice =
-            register_tag_arguments(subs, rank, pools, arena, types, stack, arguments_slice);
-
-        let new_variable_slices =
-            SubsSlice::extend_new(&mut subs.variable_slices, [variable_slice]);
-
-        macro_rules! subs_tag_name {
-            ($tag_name_slice:expr) => {
-                return UnionTags::from_slices($tag_name_slice, new_variable_slices)
-            };
-        }
-
-        match types[tags][0].0.as_str() {
-            "Ok" => subs_tag_name!(Subs::TAG_NAME_OK.as_slice()),
-            "Err" => subs_tag_name!(Subs::TAG_NAME_ERR.as_slice()),
-            "InvalidNumStr" => subs_tag_name!(Subs::TAG_NAME_INVALID_NUM_STR.as_slice()),
-            "BadUtf8" => subs_tag_name!(Subs::TAG_NAME_BAD_UTF_8.as_slice()),
-            "OutOfBounds" => subs_tag_name!(Subs::TAG_NAME_OUT_OF_BOUNDS.as_slice()),
-            _other => {}
-        }
-    }
-
-    let new_variable_slices = SubsSlice::reserve_variable_slices(subs, tags.len());
-    match find_tag_name_run(&types[tags], subs) {
-        Some(new_tag_names) => {
-            let it = (new_variable_slices.indices()).zip(payload_slices.into_iter());
-
-            for (variable_slice_index, arguments_index) in it {
-                let arguments = types[arguments_index];
-                subs.variable_slices[variable_slice_index] =
-                    register_tag_arguments(subs, rank, pools, arena, types, stack, arguments);
-            }
-
-            UnionTags::from_slices(new_tag_names, new_variable_slices)
-        }
-        None => {
-            let new_tag_names = SubsSlice::reserve_tag_names(subs, tags.len());
-
-            let it = (new_variable_slices.indices())
-                .zip(new_tag_names.indices())
-                .zip(tags.into_iter())
-                .zip(payload_slices.into_iter());
-
-            for (((variable_slice_index, tag_name_index), tag_name), arguments_index) in it {
-                let arguments = types[arguments_index];
-                subs.variable_slices[variable_slice_index] =
-                    register_tag_arguments(subs, rank, pools, arena, types, stack, arguments);
-
-                subs.tag_names[tag_name_index] = types[tag_name].clone();
-            }
-
-            UnionTags::from_slices(new_tag_names, new_variable_slices)
-        }
-    }
-}
-
-fn insert_tags_slow_path(
-    subs: &mut Subs,
-    rank: Rank,
-    pools: &mut Pools,
-    arena: &'_ bumpalo::Bump,
-    types: &mut Types,
-    union_tags: UnionTags,
-    mut tag_vars: bumpalo::collections::Vec<(TagName, VariableSubsSlice)>,
-    stack: &mut bumpalo::collections::Vec<'_, TypeToVar>,
-) -> UnionTags {
-    let (tags, payload_slices) = types.union_tag_slices(union_tags);
-
-    for (tag_index, tag_argument_types_index) in (tags.into_iter()).zip(payload_slices.into_iter())
-    {
-        let tag_argument_types = &types[tag_argument_types_index];
-
-        let new_slice = VariableSubsSlice::reserve_into_subs(subs, tag_argument_types.len());
-
-        for (i, arg) in (new_slice.indices()).zip(tag_argument_types.into_iter()) {
-            let var = RegisterVariable::with_stack(subs, rank, pools, arena, types, arg, stack);
-            subs.variables[i] = var;
-        }
-
-        tag_vars.push((types[tag_index].clone(), new_slice));
-    }
-
-    sort_and_deduplicate(&mut tag_vars);
-
-    UnionTags::insert_slices_into_subs(subs, tag_vars)
-}
-
-fn type_to_union_tags(
-    subs: &mut Subs,
-    rank: Rank,
-    pools: &mut Pools,
-    arena: &'_ bumpalo::Bump,
-    types: &mut Types,
-    union_tags: UnionTags,
-    opt_ext_slice: Slice<TypeTag>,
-    ext_openness: ExtImplicitOpenness,
-    stack: &mut bumpalo::collections::Vec<'_, TypeToVar>,
-) -> (UnionTags, TagExt) {
-    use bumpalo::collections::Vec;
-
-    let (tags, _) = types.union_tag_slices(union_tags);
-
-    let sorted = tags.len() == 1 || sorted_no_duplicate_tags(&types[tags]);
-
-    debug_assert!(opt_ext_slice.len() <= 1);
-
-    match opt_ext_slice.into_iter().next() {
-        None => {
-            let ext = Variable::EMPTY_TAG_UNION;
-
-            let union_tags = if sorted {
-                insert_tags_fast_path(subs, rank, pools, arena, types, union_tags, stack)
-            } else {
-                let tag_vars = Vec::with_capacity_in(tags.len(), arena);
-                insert_tags_slow_path(subs, rank, pools, arena, types, union_tags, tag_vars, stack)
-            };
-
-            (union_tags, TagExt::Any(ext))
-        }
-        Some(ext) => {
-            let mut tag_vars = Vec::with_capacity_in(tags.len(), arena);
-
-            let temp_ext = {
-                let temp_ext_var =
-                    RegisterVariable::with_stack(subs, rank, pools, arena, types, ext, stack);
-                TagExt::from_can(temp_ext_var, ext_openness)
-            };
-            let (it, ext) =
-                roc_types::types::gather_tags_unsorted_iter(subs, UnionTags::default(), temp_ext)
-                    .expect("extension var could not be seen as tag union");
-
-            tag_vars.extend(it.map(|(n, v)| (n.clone(), v)));
-
-            let union_tags = if tag_vars.is_empty() && sorted {
-                insert_tags_fast_path(subs, rank, pools, arena, types, union_tags, stack)
-            } else {
-                insert_tags_slow_path(subs, rank, pools, arena, types, union_tags, tag_vars, stack)
-            };
-
-            (union_tags, ext)
-        }
-    }
-}
-
-fn create_union_lambda(
-    subs: &mut Subs,
-    rank: Rank,
-    pools: &mut Pools,
-    arena: &'_ bumpalo::Bump,
-    types: &mut Types,
-    closure: Symbol,
-    capture_types: Slice<TypeTag>,
-    stack: &mut bumpalo::collections::Vec<'_, TypeToVar>,
-) -> UnionLambdas {
-    let variable_slice =
-        register_tag_arguments(subs, rank, pools, arena, types, stack, capture_types);
-    let new_variable_slices = SubsSlice::extend_new(&mut subs.variable_slices, [variable_slice]);
-
-    let lambda_name_slice = SubsSlice::extend_new(&mut subs.symbol_names, [closure]);
-
-    UnionLambdas::from_slices(lambda_name_slice, new_variable_slices)
-}
-
 fn check_for_infinite_type(
     subs: &mut Subs,
     pools: &mut Pools,
@@ -4064,370 +2383,13 @@ pub(crate) fn introduce(subs: &mut Subs, rank: Rank, pools: &mut Pools, vars: &[
     pool.extend(vars);
 }
 
-pub(crate) fn deep_copy_var_in(
+#[inline(always)]
+pub(crate) fn register(
     subs: &mut Subs,
     rank: Rank,
     pools: &mut Pools,
-    var: Variable,
-    arena: &Bump,
+    content: Content,
 ) -> Variable {
-    let mut visited = bumpalo::collections::Vec::with_capacity_in(256, arena);
-
-    let pool = pools.get_mut(rank);
-
-    let var = subs.get_root_key(var);
-    match deep_copy_var_decision(subs, rank, var) {
-        ControlFlow::Break(copy) => copy,
-        ControlFlow::Continue(copy) => {
-            deep_copy_var_help(subs, rank, pool, &mut visited, var, copy);
-
-            // we have tracked all visited variables, and can now traverse them
-            // in one go (without looking at the UnificationTable) and clear the copy field
-            for var in visited {
-                subs.set_copy_unchecked(var, OptVariable::NONE);
-            }
-
-            copy
-        }
-    }
-}
-
-#[inline]
-fn has_trivial_copy(subs: &Subs, root_var: Variable) -> Option<Variable> {
-    let existing_copy = subs.get_copy_unchecked(root_var);
-
-    if let Some(copy) = existing_copy.into_variable() {
-        Some(copy)
-    } else if subs.get_rank_unchecked(root_var) != Rank::GENERALIZED {
-        Some(root_var)
-    } else {
-        None
-    }
-}
-
-#[inline]
-fn deep_copy_var_decision(
-    subs: &mut Subs,
-    max_rank: Rank,
-    var: Variable,
-) -> ControlFlow<Variable, Variable> {
-    let var = subs.get_root_key(var);
-    if let Some(copy) = has_trivial_copy(subs, var) {
-        ControlFlow::Break(copy)
-    } else {
-        let copy_descriptor = Descriptor {
-            content: Content::Structure(FlatType::EmptyTagUnion),
-            rank: max_rank,
-            mark: Mark::NONE,
-            copy: OptVariable::NONE,
-        };
-
-        let copy = subs.fresh(copy_descriptor);
-
-        // Link the original variable to the new variable. This lets us
-        // avoid making multiple copies of the variable we are instantiating.
-        //
-        // Need to do this before recursively copying to avoid looping.
-        subs.set_mark_unchecked(var, Mark::NONE);
-        subs.set_copy_unchecked(var, copy.into());
-
-        ControlFlow::Continue(copy)
-    }
-}
-
-fn deep_copy_var_help(
-    subs: &mut Subs,
-    max_rank: Rank,
-    pool: &mut Vec<Variable>,
-    visited: &mut bumpalo::collections::Vec<'_, Variable>,
-    initial_source: Variable,
-    initial_copy: Variable,
-) -> Variable {
-    use roc_types::subs::Content::*;
-    use roc_types::subs::FlatType::*;
-
-    struct DeepCopyVarWork {
-        source: Variable,
-        copy: Variable,
-    }
-
-    let initial = DeepCopyVarWork {
-        source: initial_source,
-        copy: initial_copy,
-    };
-    let mut stack = vec![initial];
-
-    macro_rules! work {
-        ($variable:expr) => {{
-            let var = subs.get_root_key($variable);
-            match deep_copy_var_decision(subs, max_rank, var) {
-                ControlFlow::Break(copy) => copy,
-                ControlFlow::Continue(copy) => {
-                    stack.push(DeepCopyVarWork { source: var, copy });
-
-                    copy
-                }
-            }
-        }};
-    }
-
-    macro_rules! copy_sequence {
-        ($length:expr, $variables:expr) => {{
-            let new_variables = SubsSlice::reserve_into_subs(subs, $length as _);
-            for (target_index, var_index) in (new_variables.indices()).zip($variables) {
-                let var = subs[var_index];
-                let copy_var = work!(var);
-                subs.variables[target_index] = copy_var;
-            }
-
-            new_variables
-        }};
-    }
-
-    macro_rules! copy_union {
-        ($tags:expr) => {{
-            let new_variable_slices = SubsSlice::reserve_variable_slices(subs, $tags.len());
-
-            let it = (new_variable_slices.indices()).zip($tags.variables());
-            for (target_index, index) in it {
-                let slice = subs[index];
-
-                let new_variables = copy_sequence!(slice.len(), slice);
-                subs.variable_slices[target_index] = new_variables;
-            }
-
-            UnionLabels::from_slices($tags.labels(), new_variable_slices)
-        }};
-    }
-
-    // When generalizing annotations with `Openness` extensions
-    // we want to promote them to `Any`, so that usages at
-    // specialized sites can grow unboundedly and are not bound to
-    // openness-polymorphism.
-    macro_rules! copy_tag_ext {
-        ($ext:expr) => {
-            TagExt::Any(work!($ext.var()))
-        };
-    }
-
-    while let Some(DeepCopyVarWork { source: var, copy }) = stack.pop() {
-        visited.push(var);
-        pool.push(copy);
-
-        let content = *subs.get_content_unchecked(var);
-
-        // Now we recursively copy the content of the variable.
-        // We have already marked the variable as copied, so we
-        // will not repeat this work or crawl this variable again.
-        match content {
-            Structure(flat_type) => {
-                let new_flat_type = match flat_type {
-                    Apply(symbol, arguments) => {
-                        let new_arguments = copy_sequence!(arguments.len(), arguments);
-
-                        Apply(symbol, new_arguments)
-                    }
-
-                    Func(arguments, closure_var, ret_var) => {
-                        let new_ret_var = work!(ret_var);
-                        let new_closure_var = work!(closure_var);
-
-                        let new_arguments = copy_sequence!(arguments.len(), arguments);
-
-                        Func(new_arguments, new_closure_var, new_ret_var)
-                    }
-
-                    same @ EmptyRecord | same @ EmptyTuple | same @ EmptyTagUnion => same,
-
-                    Record(fields, ext_var) => {
-                        let record_fields = {
-                            let new_variables =
-                                copy_sequence!(fields.len(), fields.iter_variables());
-
-                            // When copying a let-generalized record to a specialized region, rigid
-                            // optionals just become optionals.
-                            let field_types = subs.get_subs_slice(fields.record_fields());
-                            let has_rigid_optional_field = field_types
-                                .iter()
-                                .any(|f| matches!(f, RecordField::RigidOptional(..)));
-
-                            let new_field_types_start = if has_rigid_optional_field {
-                                let field_types = field_types.to_vec();
-                                let slice = SubsSlice::extend_new(
-                                    &mut subs.record_fields,
-                                    field_types.into_iter().map(|f| match f {
-                                        RecordField::RigidOptional(())
-                                        | RecordField::RigidRequired(()) => internal_error!("Rigid optional/required should be generalized to non-rigid by this point"),
-
-                                        RecordField::Demanded(_)
-                                        | RecordField::Required(_)
-                                        | RecordField::Optional(_) => f,
-                                    }),
-                                );
-                                slice.start
-                            } else {
-                                fields.field_types_start
-                            };
-
-                            RecordFields {
-                                length: fields.length,
-                                field_names_start: fields.field_names_start,
-                                variables_start: new_variables.start,
-                                field_types_start: new_field_types_start,
-                            }
-                        };
-
-                        Record(record_fields, work!(ext_var))
-                    }
-
-                    Tuple(elems, ext_var) => {
-                        let tuple_elems = {
-                            let new_variables = copy_sequence!(elems.len(), elems.iter_variables());
-
-                            TupleElems {
-                                length: elems.length,
-                                variables_start: new_variables.start,
-                                elem_index_start: elems.elem_index_start,
-                            }
-                        };
-
-                        Tuple(tuple_elems, work!(ext_var))
-                    }
-
-                    TagUnion(tags, ext_var) => {
-                        let union_tags = copy_union!(tags);
-
-                        TagUnion(union_tags, copy_tag_ext!(ext_var))
-                    }
-
-                    FunctionOrTagUnion(tag_name, symbol, ext_var) => {
-                        FunctionOrTagUnion(tag_name, symbol, copy_tag_ext!(ext_var))
-                    }
-
-                    RecursiveTagUnion(rec_var, tags, ext_var) => {
-                        let union_tags = copy_union!(tags);
-
-                        RecursiveTagUnion(work!(rec_var), union_tags, copy_tag_ext!(ext_var))
-                    }
-                };
-
-                subs.set_content_unchecked(copy, Structure(new_flat_type));
-            }
-
-            FlexVar(_) | FlexAbleVar(_, _) | Error => {
-                subs.set_content_unchecked(copy, content);
-            }
-
-            RecursionVar {
-                opt_name,
-                structure,
-            } => {
-                let content = RecursionVar {
-                    opt_name,
-                    structure: work!(structure),
-                };
-
-                subs.set_content_unchecked(copy, content);
-            }
-
-            RigidVar(name) => {
-                subs.set_content_unchecked(copy, FlexVar(Some(name)));
-            }
-
-            RigidAbleVar(name, ability) => {
-                subs.set_content_unchecked(copy, FlexAbleVar(Some(name), ability));
-            }
-
-            Alias(symbol, arguments, real_type_var, kind) => {
-                let new_variables =
-                    copy_sequence!(arguments.all_variables_len, arguments.all_variables());
-
-                let new_arguments = AliasVariables {
-                    variables_start: new_variables.start,
-                    ..arguments
-                };
-
-                let new_real_type_var = work!(real_type_var);
-                let new_content = Alias(symbol, new_arguments, new_real_type_var, kind);
-
-                subs.set_content_unchecked(copy, new_content);
-            }
-
-            LambdaSet(subs::LambdaSet {
-                solved,
-                recursion_var,
-                unspecialized,
-                ambient_function: ambient_function_var,
-            }) => {
-                let lambda_set_var = copy;
-
-                let new_solved = copy_union!(solved);
-                let new_rec_var = recursion_var.map(|v| work!(v));
-                let new_unspecialized = SubsSlice::reserve_uls_slice(subs, unspecialized.len());
-
-                for (new_uls_index, uls_index) in
-                    (new_unspecialized.into_iter()).zip(unspecialized.into_iter())
-                {
-                    let Uls(var, sym, region) = subs[uls_index];
-                    let new_var = work!(var);
-
-                    deep_copy_uls_precondition(subs, var, new_var);
-
-                    subs[new_uls_index] = Uls(new_var, sym, region);
-
-                    subs.uls_of_var.add(new_var, lambda_set_var);
-                }
-
-                let new_ambient_function_var = work!(ambient_function_var);
-                debug_assert_ne!(
-                    ambient_function_var, new_ambient_function_var,
-                    "lambda set cloned but its ambient function wasn't?"
-                );
-
-                subs.set_content_unchecked(
-                    lambda_set_var,
-                    LambdaSet(subs::LambdaSet {
-                        solved: new_solved,
-                        recursion_var: new_rec_var,
-                        unspecialized: new_unspecialized,
-                        ambient_function: new_ambient_function_var,
-                    }),
-                );
-            }
-
-            RangedNumber(range) => {
-                let new_content = RangedNumber(range);
-
-                subs.set_content_unchecked(copy, new_content);
-            }
-        }
-    }
-
-    initial_copy
-}
-
-#[inline(always)]
-fn deep_copy_uls_precondition(subs: &Subs, original_var: Variable, new_var: Variable) {
-    if cfg!(debug_assertions) {
-        let content = subs.get_content_without_compacting(original_var);
-
-        debug_assert!(
-            matches!(
-                content,
-                Content::FlexAbleVar(..) | Content::RigidAbleVar(..)
-            ),
-            "var in unspecialized lamba set is not bound to an ability, it is {:?}",
-            roc_types::subs::SubsFmtContent(content, subs)
-        );
-        debug_assert!(
-            original_var != new_var,
-            "unspecialized lamba set var was not instantiated"
-        );
-    }
-}
-
-#[inline(always)]
-fn register(subs: &mut Subs, rank: Rank, pools: &mut Pools, content: Content) -> Variable {
     let descriptor = Descriptor {
         content,
         rank,
@@ -4442,7 +2404,7 @@ fn register(subs: &mut Subs, rank: Rank, pools: &mut Pools, content: Content) ->
     var
 }
 
-fn register_with_known_var(
+pub(crate) fn register_with_known_var(
     subs: &mut Subs,
     var: Variable,
     rank: Rank,
